[TD-578] 전반적인 코드 정리
