+++ threedollar-admin/src/integrationTest/kotlin/com/depromeet/threedollar/admin/controller/HealthCheckControllerTest.kt
@@ -1,25 +1,25 @@
-package com.depromeet.threedollar.api.controller
+package com.depromeet.threedollar.admin.controller
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.TestConstructor
 import org.springframework.test.web.servlet.MockMvc
+import org.springframework.test.web.servlet.get
 
-import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get
-import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
-
-@AutoConfigureMockMvc
 @TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
+@AutoConfigureMockMvc
 @SpringBootTest
-class HealthControllerTest(
+class HealthCheckControllerTest(
     private val mockMvc: MockMvc
 ) {
 
     @Test
     fun healthCheck() {
-        this.mockMvc.perform(get("/ping"))
-            .andExpect(status().isOk)
+        mockMvc.get("/ping")
+            .andExpect {
+                status { isOk() }
+            }
     }
 
 }
