+++ threedollar-admin/src/main/kotlin/com/depromeet/threedollar/admin/controller/HealthCheckController.kt
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
-class HealthController {
+class HealthCheckController {
 
     @GetMapping("/", "/ping")
     fun healthCheck(): ApiResponse<String> {
