+++ threedollar-api-user/src/integrationTest/java/com/depromeet/threedollar/api/controller/HealthCheckControllerTest.kt
@@ -1,25 +1,25 @@
-package com.depromeet.threedollar.admin.controller
+package com.depromeet.threedollar.api.controller
 
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.TestConstructor
 import org.springframework.test.web.servlet.MockMvc
-import org.springframework.test.web.servlet.get
 
-@TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
+
 @AutoConfigureMockMvc
+@TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
 @SpringBootTest
-class HealthControllerTest(
+class HealthCheckControllerTest(
     private val mockMvc: MockMvc
 ) {
 
     @Test
     fun healthCheck() {
-        mockMvc.get("/ping")
-            .andExpect {
-                status { isOk() }
-            }
+        this.mockMvc.perform(get("/ping"))
+            .andExpect(status().isOk)
     }
 
 }
