+++ threedollar-api-user/src/main/java/com/depromeet/threedollar/api/controller/HealthCheckController.kt
@@ -1,4 +1,4 @@
-package com.depromeet.threedollar.boss.api.controller
+package com.depromeet.threedollar.api.controller
 
 import com.depromeet.threedollar.application.common.dto.ApiResponse
 import io.swagger.annotations.ApiOperation
@@ -6,12 +6,12 @@ import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
-class MainController {
+class HealthCheckController {
 
     @ApiOperation("Health Check")
-    @GetMapping("/", "/ping")
-    fun ping(): ApiResponse<String> {
-        return ApiResponse.success("가슴속 3천원 사장님 API 서버")
+    @GetMapping(value = ["/", "/ping"])
+    fun healthCheck(): ApiResponse<String> {
+        return ApiResponse.success("가슴속 3천원 유저 API 서버")
     }
 
 }
