+++ threedollar-domain/src/main/java/com/depromeet/threedollar/domain/user/domain/medal/UserMedalStatus.java
@@ -4,6 +4,5 @@ public enum UserMedalStatus {
 
     ACTIVE,
     IN_ACTIVE,
-    ;
 
 }
